OM-42837: enable ON-Demand report (VM group summary) to support discovered and user created group (RB29872)
Intent:
=======
This is continuation on enabling classic reports in XL.
These changes are to enable ON-Demand report, which user can select a group (VM group in this case).

Detail changes:
===============
1. updated "vm group" report design file to not using "utilization" in entities table (since XL doesn't populate this property subtype)
2. enabled ON-Demand "vm group" report:
a. user select a group from UI and send request to reporting component
b. report component get the group id and ask group component for details
c. getting the group members depended on discovered group or user created group
d. populate the group member relationships.

+++ com.vmturbo.reports.component/src/test/java/com/vmturbo/reports/component/ReportingTestConfig.java
@@ -28,13 +28,6 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 import com.google.common.collect.ImmutableMap;
 
-import io.grpc.Channel;
-
-import com.vmturbo.action.orchestrator.api.impl.ActionOrchestratorClientConfig;
-import com.vmturbo.common.protobuf.action.ActionsServiceGrpc;
-import com.vmturbo.common.protobuf.action.ActionsServiceGrpc.ActionsServiceBlockingStub;
-import com.vmturbo.common.protobuf.group.GroupServiceGrpc;
-import com.vmturbo.common.protobuf.group.GroupServiceGrpc.GroupServiceBlockingStub;
 import com.vmturbo.common.protobuf.setting.SettingServiceGrpc;
 import com.vmturbo.commons.idgen.IdentityGenerator;
 import com.vmturbo.components.api.test.GrpcTestServer;
@@ -43,13 +36,11 @@ import com.vmturbo.components.common.mail.MailManager;
 import com.vmturbo.group.api.GroupClientConfig;
 import com.vmturbo.reporting.api.ReportingNotificationReceiver;
 import com.vmturbo.reporting.api.protobuf.Reporting.ReportNotification;
-import com.vmturbo.reporting.api.protobuf.ReportingREST.ReportingServiceController;
 import com.vmturbo.reporting.api.protobuf.ReportingServiceGrpc.ReportingServiceImplBase;
 import com.vmturbo.reports.component.communication.ReportNotificationSender;
 import com.vmturbo.reports.component.communication.ReportNotificationSenderImpl;
 import com.vmturbo.reports.component.communication.ReportingServiceRpc;
 import com.vmturbo.reports.component.data.GroupGeneratorDelegate;
-import com.vmturbo.reports.component.data.ReportDBDataWriter;
 import com.vmturbo.reports.component.data.ReportTemplate;
 import com.vmturbo.reports.component.data.ReportsDataContext;
 import com.vmturbo.reports.component.data.ReportsDataGenerator;
@@ -140,11 +131,11 @@ public class ReportingTestConfig {
                 mailManager(), new ReportsDataGenerator(mock(ReportsDataContext.class), getReportMap()));
     }
 
-    private Map<Integer, ReportTemplate> getReportMap(){
+    private Map<Long, ReportTemplate> getReportMap(){
         final GroupGeneratorDelegate delegate = new GroupGeneratorDelegate();
-        return ImmutableMap.of(150, new Daily_vm_rightsizing_advice_grid(delegate),
-            184, new Daily_vm_over_under_prov_grid_30_days(delegate),
-            146, new Monthly_cluster_summary(delegate));
+        return ImmutableMap.of(150L, new Daily_vm_rightsizing_advice_grid(delegate),
+            184L, new Daily_vm_over_under_prov_grid_30_days(delegate),
+            146L, new Monthly_cluster_summary(delegate));
     }
 
     @Bean
